refactor: mejoras globales de estructura, internacionalización y UI
- Separación de secciones principales en componentes reutilizables (Header, Footer, HeroSection, FeaturesStrip, WhyChooseUs, BeforeAfterGallery, Testimonials, LeadCaptureForm, LocalSeoSection).
- Implementación de contexto global de idioma (LanguageContext) para traducción dinámica en toda la landing.
- Cambio de idioma ahora afecta a todos los componentes, controlado desde Header y Footer.
- Mejora visual del switch de idioma: banderas de México y EE.UU. en Header y Footer.
- El botón “Book a Free Consultation” ahora redirige correctamente a la sección del formulario desde cualquier página.
- El logo en el Header ahora redirige a la página principal.
- Mejora visual del mapa en LocalSeoSection: bordes, sombra, leyenda y efecto hover.
- Limpieza de duplicados: Header y Footer solo se renderizan desde el layout global.
- Ajustes menores de UI y navegación para una experiencia más profesional y consistente.
